[skip ci] rename tst
