+++ src/optimum/rbln/transformers/models/time_series_transformers/__init__.py
@@ -23,4 +23,4 @@
 
 from ....ops import paged_add_softmax_attn_decode, rbln_cache_update
 from .configuration_time_series_transformer import RBLNTimeSeriesTransformerForPredictionConfig
-from .modeling_time_series_transformers import RBLNTimeSeriesTransformerForPrediction
+from .modeling_time_series_transformer import RBLNTimeSeriesTransformerForPrediction
