+++ src/optimum/rbln/transformers/models/time_series_transformers/modeling_time_series_transformer.py
